chore: move colors to another file
